Improve error message for unsupported platform
In parser.py, the UnsupportedPlatformError raised by parse() included
`platform` rather than `parsed_url.platform`, leading to a misleading
error message. This commit changes the error to use `parsed_url.host`
instead, which is more helpful for diagnosing the error. It also
changes `UnsupportedPlatformError` to take a list of supported
platforms, which is helpful to the caller in printing out a helpful
error message.