docs!: Improve error message for unsupported VCS platform (bug 1793219) (#6)
In parser.py, the UnsupportedPlatformError raised by parse() included
`platform` rather than `parsed_url.platform`, leading to a misleading
error message. This commit changes the error to use `parsed_url.host`
instead, which is more helpful for diagnosing the error. It also
changes `UnsupportedPlatformError` to take a list of supported
platforms, which is helpful to the caller in printing out a helpful
error message.

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Johan Lorenzo <jlorenzo@mozilla.com>